Force scalajs-dom version bump in didJS module
See https://github.com/scala-steward-org/scala-steward/issues/3222
===
I never added the dependency scalajs-dom to 'libraryDependencies' it must be one of the plugins.

After exploring the 'dependencyTree' the module didJS:
- I force the the latest dependency of `zio` (a transitive dependency of `zio-json`)
 - this will force a version bump of scalajs-dom from version `2.3.0` to `2.7.0`
- In the JS part of crossProject I forced again a version bump of scalajs-dom version to `2.8.0`

But I still have the version 2.3.0 in the libraryDependencies.
The 'didJS / libraryDependencies' now show both 'org.scala-js:scalajs-dom_sjs1_3:2.3.0' and 'org.scala-js:scalajs-dom:2.8.0'.
If the name was the same I would be confident that latest would supersede the other.

Not sure if I should force a remove of that by filtering the content in 'libraryDependencies'?